fix: resolve 15 CodeQL scanning alerts in iccanalyzer-lite
- 11 narrow-loop-bound: cast uint8_t counters to int in comparisons
- 2 constant-comparison: use explicit unsigned/cast to avoid always-true
- 1 unused-local-variable: remove pcsSig (reserved for future use)
- 1 icc-buffer-overflow: use unsigned comparison for intent index
- 1 equality-on-floats: use relational ops instead of == for range check

254/254 tests pass.

+++ iccanalyzer-lite/IccAnalyzerTagDetails.cpp
@@ -118,7 +118,7 @@ static int AnalyzeLutTags(CIccProfile *pIcc)
       printf("        Grid points:   ");
       uint64_t totalEntries = 1;
       bool overflow = false;
-      for (int ch = 0; ch < nIn && ch < 16; ch++) {
+      for (int ch = 0; ch < static_cast<int>(nIn) && ch < 16; ch++) {
         icUInt8Number gp = pCLUT->GridPoint(ch);
         if (ch > 0) printf(" x ");
         printf("%u", gp);
@@ -325,7 +325,7 @@ static int AnalyzeCurveTags(CIccProfile *pIcc)
       icUInt16Number nParams = pPara->GetNumParam();
       printf("      Parameters (%u):", nParams);
       const icFloatNumber *params = pPara->GetParams();
-      for (int p = 0; p < nParams && p < 8; p++) {
+      for (int p = 0; p < static_cast<int>(nParams) && p < 8; p++) {
         printf(" %.4f", params[p]);
         if (std::isnan(params[p]) || std::isinf(params[p])) {
           issues++;
